requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pyyaml==6.0.1
apscheduler==3.10.4
requests==2.31.0
websockets==12.0
